packages/convo-api-py/pyproject.toml
[tool.poetry]
name = "convo-api-py"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "convo", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.9.2"
python-magic = "^0.4.27"
langchain-community = "^0.3.1"
openai = "^1.51.2"
langchain = "^0.3.2"
psycopg = "^3.2.3"
botocore = "^1.35.36"
iyio_common = { path = "../iyio-common-py", develop=true }
boto3 = "^1.35.36"
psycopg2-binary = "^2.9.9"
pdfminer-six = "^20240706"
unstructured = "^0.15.13"
# Need to use this fork until build issues are resolved with the main age Python package
apache-age-python = { git = "https://github.com/zombie-einstein/age.git", subdirectory = "drivers/python" }
nano-graphrag = "^0.0.8"
markdown = "^3.7"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
langchain-unstructured = "^0.1.5"
databases = {version = "^0.9.0", extras = ["postgresql"]}
pgvector = "^0.3.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
